# Implement class-based dispatch for term handling