--- a/core/src/main/java/eu/neverblink/jelly/core/ProtoEncoderConverter.java
+++ b/core/src/main/java/eu/neverblink/jelly/core/ProtoEncoderConverter.java
@@ -8,6 +8,23 @@ package eu.neverblink.jelly.core;
  * @param <TNode> type of RDF nodes in the library
  */
 public interface ProtoEncoderConverter<TNode> {
+    /**
+     * Convert a subject/predicate/object node to a Jelly proto object.
+     *
+     * @param encoder encoder to use for creating Jelly proto objects. Call its methods to create
+     *                new Jelly proto objects.
+     * @param node node to convert
+     * @return Jelly proto object representing the node, obtained from the encoder.
+     */
     Object nodeToProto(NodeEncoder<TNode> encoder, TNode node);
+
+    /**
+     * Convert a graph node to a Jelly proto object.
+     *
+     * @param encoder encoder to use for creating Jelly proto objects. Call its methods to create
+     *                new Jelly proto objects.
+     * @param node graph node to convert. If null, this represents the default graph.
+     * @return Jelly proto object representing the graph node, obtained from the encoder.
+     */
     Object graphNodeToProto(NodeEncoder<TNode> encoder, TNode node);
 }