--- a/core/src/main/java/eu/neverblink/jelly/core/RdfBufferAppender.java
+++ b/core/src/main/java/eu/neverblink/jelly/core/RdfBufferAppender.java
@@ -8,13 +8,11 @@ import eu.neverblink.jelly.core.proto.v1.*;
  * This is used by NodeEncoder.
  */
 public interface RdfBufferAppender<TNode> {
-    byte TERM_IRI = 1;
-    byte TERM_BNODE = 2;
-    byte TERM_LITERAL = 3;
-    byte TERM_TRIPLE = 4;
-    byte TERM_DEFAULT_GRAPH = 5;
-    
-    record Encoded(Object node, byte termType) { }
+    //    byte TERM_IRI = 1;
+    //    byte TERM_BNODE = 2;
+    //    byte TERM_LITERAL = 3;
+    //    byte TERM_TRIPLE = 4;
+    //    byte TERM_DEFAULT_GRAPH = 5;
 
     // Lookup entries
     void appendNameEntry(RdfNameEntry nameEntry);
@@ -25,6 +23,6 @@ public interface RdfBufferAppender<TNode> {
     //    void appendIri(RdfIri iri);
     //    void appendBlankNode(String label);
     //    void appendLiteral(RdfLiteral literal);
-    Encoded appendQuotedTriple(TNode subject, TNode predicate, TNode object);
+    RdfTriple appendQuotedTriple(TNode subject, TNode predicate, TNode object);
     //    void appendDefaultGraph();
 }