… core-patch/src/test/scala/eu/neverblink/jelly/core/patch/helpers/PatchAdapter.scala
… core-patch/src/test/scala/eu/neverblink/jelly/core/patch/helpers/PatchTestCases.scala
… core/src/main/java/eu/neverblink/jelly/core/internal/ProtoDecoderImpl.java
… core/src/main/java/eu/neverblink/jelly/core/internal/ProtoTranscoderImpl.java
… core/src/main/java/eu/neverblink/jelly/core/internal/proto/GraphBase.java
… core/src/main/java/eu/neverblink/jelly/core/internal/proto/HeaderBase.java
… core/src/main/java/eu/neverblink/jelly/core/internal/proto/SpoBase.java
… core/src/test/scala/eu/neverblink/jelly/core/ProtoAuxiliarySpec.scala
… core/src/test/scala/eu/neverblink/jelly/core/helpers/MockProtoEncoderConverter.scala
… core/src/test/scala/eu/neverblink/jelly/core/helpers/RdfAdapter.scala
… core/src/test/scala/eu/neverblink/jelly/core/internal/NodeEncoderSpec.scala
… crunchy-protoc-plugin/src/main/scala/eu/neverblink/protoc/java/gen/MessageGenerator.scala
… crunchy-protoc-plugin/src/main/scala/eu/neverblink/protoc/java/gen/OneOfGenerator.scala
… crunchy-protoc-plugin/src/main/scala/eu/neverblink/protoc/java/gen/PluginOptions.scala
… crunchy-protoc-plugin/src/main/scala/eu/neverblink/protoc/java/gen/RequestInfo.scala
… jena/src/main/java/eu/neverblink/jelly/convert/jena/JenaEncoderConverter.java
… jena/src/test/scala/eu/neverblink/jelly/convert/jena/JenaProtoEncoderSpec.scala
… jmh/src/main/scala/eu/neverblink/jelly/jmh/RdfTripleRecursiveBench.scala
… rdf-protos/src/main/args.txt
… rdf4j/src/main/java/eu/neverblink/jelly/convert/rdf4j/Rdf4jEncoderConverter.java
… rdf4j/src/test/scala/eu/neverblink/jelly/convert/rdf4j/rio/JellyParserSpec.scala
… titanium-rdf-api/src/main/java/eu/neverblink/jelly/convert/titanium/internal/TitaniumEncoderConverter.java